refactor: consolidate world folder
