+++ latebit/core/world/ObjectUtils.cpp
