+++ latebit/core/world/ObjectUtils.h
