+++ latebit/core/world/Scene.test.cpp
@@ -1,5 +1,5 @@
 #include "Scene.h"
-#include "latebit/core/objects/Object.h"
+#include "latebit/core/world/Object.h"
 #include "latebit/core/world/WorldManager.h"
 #include "test/lib/test.h"
