+++ latebit/core/world/SceneGraph.test.cpp
@@ -2,8 +2,8 @@
 
 #include <memory>
 
-#include "latebit/core/objects/Object.h"
-#include "latebit/core/objects/ObjectUtils.h"
+#include "latebit/core/world/Object.h"
+#include "latebit/core/world/ObjectUtils.h"
 #include "test/lib/test.h"
 
 using namespace lb;
