+++ latebit/core/world/WorldManager.cpp
@@ -7,10 +7,10 @@
 #include "latebit/core/events/EventCollision.h"
 #include "latebit/core/events/EventOut.h"
 #include "latebit/core/geometry/Vector.h"
-#include "latebit/core/objects/Object.h"
-#include "latebit/core/objects/ObjectUtils.h"
-#include "latebit/core/utils/utils.h"
+#include "latebit/core/world/Object.h"
+#include "latebit/core/world/ObjectUtils.h"
 #include "latebit/core/world/View.h"
+#include "latebit/core/utils/utils.h"
 #include "latebit/utils/Logger.h"
 
 #define WM lb::WorldManager::getInstance()
