+++ latebit/ui/ViewObject.cpp
@@ -4,7 +4,7 @@
 #include "latebit/core/geometry/Box.h"
 #include "latebit/core/geometry/Vector.h"
 #include "latebit/core/graphics/DisplayManager.h"
-#include "latebit/core/objects/SceneGraph.h"
+#include "latebit/core/world/SceneGraph.h"
 #include "latebit/core/world/WorldManager.h"
 #include "latebit/utils/Math.h"
